requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic_settings
google-generativeai
sqlmodel